refactor: service層を責務ごとにファイル分割
staywatch.goの肥大化（605行）を解消し、責務ごとに適切なファイルに分割。

変更内容:
- activity.go を作成: アクティビティ分析関連の関数を移動
  - getActivityProbability
  - getActivityTimeRange
  - getUserActivityEventIDs
  - filterByCommonActivities
  - getUserActivityTags

- notification.go を作成: 通知・メッセージフォーマット関連を移動
  - NotifyByEvent
  - findPredictionForUser
  - formatUserLine

- occupancy.go を作成: 占有率計算関連を移動
  - calculateRecommendedTimeRanges
  - findOverlappingRanges

- lib/math_utils.go を作成: 汎用的な数学関数を移動
  - Max, Min

- staywatch.go: StayWatch API関連のみに絞り込み（132行に削減）
  - URL構築ヘルパー
  - API呼び出し関数
  - 確率フィルタリング

効果:
- 各ファイルの責務が明確化
- staywatch.goのサイズを78%削減（605行 → 132行）
- 関心の分離により保守性が向上
- テスト可能性の向上
