+++ src/lib/math_utils.go
@@ -1,0 +1,17 @@
+package lib
+
+// Max returns the larger of two integers
+func Max(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+// Min returns the smaller of two integers
+func Min(a, b int) int {
+	if a < b {
+		return a
+	}
+	return b
+}
